+++ api4jenkins/mix.py
@@ -2,6 +2,8 @@
 
 # pylint: disable=no-member
 
+from collections import namedtuple
+
 
 class DeletionMixIn:
 
@@ -45,3 +47,25 @@ class RawJsonMixIn:
 
     def api_json(self, tree='', depth=0):
         return self.raw
+
+
+Parameter = namedtuple('Parameter', ['class_name', 'name', 'value'])
+
+
+class CauseMixIn:
+
+    def get_parameters(self):
+        parameters = []
+        for action in self.api_json()['actions']:
+            if 'parameters' in action:
+                for raw in action['parameters']:
+                    parameters.append(Parameter(raw['_class'], raw['name'],
+                                                raw.get('value', '')))
+                break
+        return parameters
+
+    def get_causes(self):
+        for action in self.api_json()['actions']:
+            if 'causes' in action:
+                return action['causes']
+        return []